update(server): I separated the functions into modules. tbm I added the response and error module to follow the functions that were together in utils.rs.  that were together in utils.rs. finally move all the error log messages to the logger module
